The emulator now support more environmental variables for manage 4 stages.
It is now possible to set name, probability of distribution, latitude, longitude and sigma values of the covariant matrix for 4 stages through environmental variables.
It is possible to decide if using the "docker-mode" (using environmental variable) or the "debug mode" using tha values of "Settings" class changing the value of PermanentSettings.dubug from True to False.

Main updates of this commit:
- new constants related to environmental variables were added
- bug-fixing: added "restart: on-failure" to SCRAL container to be sure to have it always on;
- added a new docker-compose-v1.0.yml to have a stable simple release of the emulator;
- docker-compose.yml should reference the last emulator version;
- some debug statements were changed to improve readability;